calculator.py
def calculate_optimal_troops(opponent_troops):
    """
    Calculates the optimal troop composition to counter the opponent's troops.

    Args:
        opponent_troops (dict): A dictionary containing the opponent's troop
                                composition.
            Example: {'bruisers': 100, 'hitmen': 50, 'bikers': 75}

    Returns:
        dict: A dictionary containing the optimal troop composition.
    """
    optimal_troops = {'bruisers': 0, 'hitmen': 0, 'bikers': 0}

    # Counter logic:
    # Bikers are strong against Bruisers
    # Bruisers are strong against Hitmen
    # Hitmen are strong against Bikers
    if 'bruisers' in opponent_troops:
        optimal_troops['bikers'] += opponent_troops['bruisers']
    if 'hitmen' in opponent_troops:
        optimal_troops['bruisers'] += opponent_troops['hitmen']
    if 'bikers' in opponent_troops:
        optimal_troops['hitmen'] += opponent_troops['bikers']

    return optimal_troops


def calculate_optimal_enforcers(user_enforcers, opponent_enforcers):
    """
    Calculates the optimal enforcer setup to counter the opponent's enforcers.

    Args:
        user_enforcers (str): A string containing the user's enforcers.
        opponent_enforcers (str): A string containing the opponent's enforcers.

    Returns:
        dict: A dictionary containing the optimal enforcer setup.
    """
    # This is a placeholder implementation. A more advanced implementation would
    # consider enforcer skills, tiers, and other factors.
    return {'optimal_enforcers': 'Placeholder'}


def calculate_resources(resources):
    """
    Calculates the total value of the user's resources.

    Args:
        resources (dict): A dictionary containing the user's resources.

    Returns:
        dict: A dictionary containing the total value of the resources.
    """
    # This is a placeholder implementation. A more advanced implementation would
    # consider item values and other factors.
    return {'total_value': 'Placeholder'}


import pytesseract
from PIL import Image
import re


def analyze_screenshot(filepath):
    """
    Analyzes a screenshot to extract game data using OCR.

    Args:
        filepath (str): The path to the screenshot file.

    Returns:
        dict: A dictionary containing the extracted data.
    """
    try:
        image = Image.open(filepath)
        text = pytesseract.image_to_string(image)

        # Basic parsing logic (highly dependent on screenshot format)
        extracted_data = {}

        # Example: Extracting troop counts
        bruisers_match = re.search(r'Bruisers: (\d+)', text)
        if bruisers_match:
            extracted_data['bruisers'] = int(bruisers_match.group(1))

        hitmen_match = re.search(r'Hitmen: (\d+)', text)
        if hitmen_match:
            extracted_data['hitmen'] = int(hitmen_match.group(1))

        bikers_match = re.search(r'Bikers: (\d+)', text)
        if bikers_match:
            extracted_data['bikers'] = int(bikers_match.group(1))

        # Example: Extracting enforcer names
        enforcers_match = re.search(r'Enforcers: (.*)', text)
        if enforcers_match:
            enforcers = [e.strip() for e in enforcers_match.group(1).split(',')]
            extracted_data['enforcers'] = enforcers

        if not extracted_data:
            return {'error': 'Could not extract any data from the screenshot.'}

        return extracted_data
    except Exception as e:
        return {'error': f'An error occurred during OCR: {str(e)}'}
