requirements.txt
bandit==1.7.10
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
flake8==7.3.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
packaging==25.0
pbr==6.1.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytesseract==0.3.13
PyYAML==6.0.2
rich==14.0.0
setuptools==80.9.0
SQLAlchemy==2.0.41
stevedore==5.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
gevent
pytesseract
pillow
